requirements_win.txt
colorama==0.4.6
filelock==3.19.1
fsspec==2025.9.0
Jinja2==3.1.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
pillow==11.3.0
pycountry==24.6.1
setuptools==70.2.0
sympy==1.14.0
torch==2.9.0+cu126
torchvision==0.24.0+cu126
tqdm==4.67.1
typing_extensions==4.15.0
